facturas 2

+++ ProyectoGestionDistribucion/src/model/FilaFacturaProveedor.java
@@ -2,13 +2,8 @@ package model;
 
 import java.io.Serializable;
 import javax.persistence.*;
-import java.math.BigDecimal;
 
 
-/**
- * The persistent class for the FILAS_FACTURAS_PROVEEDOR database table.
- * 
- */
 @Entity
 @Table(name="FILAS_FACTURAS_PROVEEDOR")
 @NamedQuery(name="FilaFacturasroveedor.findAll", query="SELECT f FROM FilaFacturaProveedor f")
@@ -22,7 +17,7 @@ public class FilaFacturaProveedor implements Serializable {
 	private int cantidad;
 
 	@Column(name="PRECIO")
-	private BigDecimal precio;
+	private double precio;
 
 	//bi-directional many-to-one association to FacturasProveedor
 	@ManyToOne
@@ -53,11 +48,11 @@ public class FilaFacturaProveedor implements Serializable {
 		this.cantidad = cantidad;
 	}
 
-	public BigDecimal getPrecio() {
+	public double getPrecio() {
 		return this.precio;
 	}
 
-	public void setPrecio(BigDecimal precio) {
+	public void setPrecio(double precio) {
 		this.precio = precio;
 	}
